% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatreat.R
\name{asciitonumber}
\alias{asciitonumber}
\title{Convert pedigree names to numeric indices using C++}
\usage{
asciitonumber(pedigree.data, unk = 0)
}
\arguments{
\item{pedigree.data}{A 3-column matrix or data frame with individual, sire, and dam}

\item{unk}{Code for unknown parent (default = "0")}
}
\value{
A list with sire, dam, and individual names
}
\description{
Convert pedigree names to numeric indices using C++
}
\author{
Rodrigo R. Amadeu \email{rramadeu@gmail.com}

Thiago de Paula Oliveira
}
