% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converttofrequency.R
\name{converttofrequency}
\alias{converttofrequency}
\title{Converts molecular (AA,AB,BB) data in a frequency format (0,0.5,1).}
\usage{
converttofrequency(file = NULL, ploidy = 2, format = 4, unk = NA,
  genotype = NULL, dominant = FALSE, output = "convdata",
  transpose = FALSE)
}
\arguments{
\item{file}{path of your file (marker in rows and individual in columns).}

\item{ploidy}{ploidy of your data 2 or 4, Default=2.}

\item{format}{1 if the data code as 0,1,2,..; 2 if if the data code as ...,-1,0,1,...; 3 if if the data code as "BB","AB","AA" or "BBBB", "ABBB", "AABB", "AAAB", "AAAA"; 4 if different. Than, you need to specify in genotype parameter.}

\item{unk}{unknown value assumed (default=NA)}

\item{genotype}{available if format=4. Please insert your genotype categories here as a list. e.g.:  genotype=c("CCCC","CGGG","CCGG","CGGG","GGGG") in this example CCCC will be coded as 0, CGGG as 0.25, CCGG as 0.5, CGGG as 0.75, G as 1. Default=NULL.}

\item{output}{choose a the name of output file. Default="convdata".}
}
\value{
csv file markers x individuals with frequencies instead of molecular code.
}
\description{
Converts molecular data in frequency format. Molecular data can be coded with numbers, letters or any ascii characters.
}
\examples{
converttofrequency()

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}

