% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatreat.R
\name{datatreat}
\alias{datatreat}
\title{Organizes pedigree data in a chronological way using C++}
\usage{
datatreat(data = NULL, n.max = 50, unk = 0, save = FALSE)
}
\arguments{
\item{data}{A 3-column data frame or matrix with individual, sire, and dam}

\item{n.max}{Max number of iterations (default = 50)}

\item{unk}{Code for unknown parent (default = 0)}

\item{save}{If TRUE, saves the reordered pedigree to a text file}
}
\value{
A list with elements: sire (numeric), dire (numeric), ind_data (names)
}
\description{
This function organizes pedigree data in a chronological way and returns a list:
i) parental 1 values (numeric); ii) parental 2 values (numeric); iii) individual names.
Also saves a `.txt` file if `save = TRUE`.
}
\examples{
# data(ped.mrode)
# datatreat(ped.mrode)
}
\author{
Rodrigo R. Amadeu \email{rramadeu@gmail.com}

Thiago de Paula Oliveira
}
