% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.matrix.R
\name{explore.matrix}
\alias{explore.matrix}
\title{Explore a relationship matrix}
\usage{
explore.matrix(data = NULL, type = NULL, print = TRUE, w = 0,
  name = NULL)
}
\arguments{
\item{data}{output of function 'Amatrix' or 'Gmatrix' or 'Hmatrix'.}

\item{type}{"A" or "G" or "H". Default=NULL.}

\item{print}{if TRUE, print analysis. Default=TRUE.}

\item{w}{proportion of parental gametas IBD due to double reduction. Default=0.}

\item{name}{csv file name}
}
\value{
list with exploratory analysis about the matrix: summary off-diagonal, summary diagonal, sort data (top shared genotypes)...
}
\description{
Given a data from 'Amatrix', 'Gmatrix', or 'Hmatrix' functions, return a list with exploratory analysis of the matrix.
}
\examples{
explore.matrix()

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}

