% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterpedigree.R
\name{filterpedigree}
\alias{filterpedigree}
\title{Filter the pedigree to keep only the genealogy of a subset of individuals}
\usage{
filterpedigree(inds, data)
}
\arguments{
\item{inds}{vector with strings of individuals to keep their genealogy in the matrix}

\item{data}{name of the pedigree data frame. Default=NULL.}
}
\value{
a data frame with pedigree containing the genealogy of the selected individuals
}
\description{
Filter the pedigree to keep only the genealogy of a subset of individuals
}
\examples{
data(ped.sol)
new.ped.sol = filterpedigree(inds = c("MSW168-2","W14090-3","W14090-4"),data=ped.sol)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
