% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{H_martini_blocks}
\alias{H_martini_blocks}
\title{Compute Martini block corrections (H11, H12, H21, H22 - A22)}
\usage{
H_martini_blocks(A12, A22, G22, tau, omega)
}
\arguments{
\item{A12}{n1 x n2 block from A}

\item{A22}{n2 x n2 block from A (genotyped)}

\item{G22}{n2 x n2 block from G (genotyped)}

\item{tau}{scalar}

\item{omega}{scalar}
}
\value{
list(H11, H12, H21, H22corr) where H22corr = H22 - A22
}
\description{
Compute Martini block corrections (H11, H12, H21, H22 - A22)
}
