% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hmatrix.R
\name{Hmatrix}
\alias{Hmatrix}
\title{Construction of Relationship Matrix H}
\usage{
Hmatrix(A = NULL, G = NULL, markers = NULL, c = 0, explore = FALSE)
}
\arguments{
\item{A}{A matrix from function Amatrix}

\item{G}{G matrix from function Gmatrix}

\item{c}{constant value of H computation}

\item{explore}{if TRUE performs exploratory analysis of the matrix}
}
\value{
H Matrix with the relationship between the individuals based on pedigree and corrected by molecular information
}
\description{
Given a matrix A and a matrix G returns a H matrix - A matrix corrected by G.
}
\examples{
Hmatrix()

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
\references{
Patricio Munoz ?
}

