% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingdata.R
\name{missingdata}
\alias{missingdata}
\title{Survying on missing data}
\usage{
missingdata(data, unk = 0)
}
\arguments{
\item{data}{data name from a pedigree list}

\item{unk}{unknown value of your data}
}
\value{
list with $conflict: rows of the data which are at least one parental name equal to the individual. $missing.sire: rows of the data which arie missing data sire (Parental 1) information. $missing.dire: same as above for dire (Parental 2). $summary.missing: summary of the missing data. 2 columns, 1st for the name of the parental listed, 2nd for the how many times appeared in the data.
}
\description{
This function verify which rows in a pedigree data has missing parental or conflictuos data
}
\examples{
data(ped.mrode)
missingdata(ped.mrode)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
