% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{munoz_var_mean_by_Aclass}
\alias{munoz_var_mean_by_Aclass}
\title{Compute var/mean of G by classes of rounded A, and fill varA/meanG}
\usage{
munoz_var_mean_by_Aclass(G, A, round_digits)
}
\arguments{
\item{G}{square matrix}

\item{A}{square matrix (same dim as G)}

\item{round_digits}{integer number of decimal places to round A}
}
\value{
list(varA, meanG, classes, var_by_class, mean_by_class)
}
\description{
Compute var/mean of G by classes of rounded A, and fill varA/meanG
}
