% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped.mrode-data.R
\docType{data}
\name{ped.mrode}
\alias{ped.mrode}
\title{Pedigree Data}
\format{table}
\usage{
data(ped.mrode)
}
\description{
Data from pedigree example proposed by Mrode 2005
}
\examples{
data(pedigree.mrode)

}
\references{
R. A. Mrode, R. Thompson. Linear Models for the Prediction of Animal Breeding Values. CABI, 2005.
}
\keyword{datasets}
