% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped.potato-data.R
\docType{data}
\name{ped.potato}
\alias{ped.potato}
\title{Autotetraploid Pedigree Data}
\format{table}
\usage{
data(ped.potato)
}
\description{
Data from pedigree derived from the Supplementary data (FileS3.csv) from Endelman (2018).
}
\examples{
data(ped.potato)

}
\references{
Endelman, J. B., et al., 2018. Genetic variance partitioning and genome-wide prediction with allele dosage information in autotetraploid potato. Genetics, 209(1) pp. 77-87.
}
\keyword{datasets}
