% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped.sol-data.R
\docType{data}
\name{ped.sol}
\alias{ped.sol}
\title{Pedigree data for autopolyploid examples}
\format{data.frame}
\usage{
data(ped.sol)
}
\description{
Dataset extract from  supplementary material from Endelman et al. (2018). 
Pedigree data frame of Potato population, missing data as 0.
}
\examples{
data(ped.sol)

}
\references{
Endelman, JB, et al., 2018 Genetic variance partitioning and genome-wide prediction with allele dosage information in autotetraploid potato. Genetics, 209(1) pp. 77-87.
}
\keyword{datasets}
