% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.genealogy.R
\name{plot.genealogy}
\alias{plot.genealogy}
\title{Construction of Genealogy for an individual}
\usage{
\method{plot}{genealogy}(data = data, ind = NULL, gen = 20, Amat = NULL,
  fontsize = 3, save.image = TRUE, return.plot = TRUE, angle.names = 0,
  big = FALSE)
}
\arguments{
\item{data}{pedigree data name (data frame in a 3-column way format).}

\item{ind}{indiviudal name (string)}

\item{gen}{the max number of generations ago of the plot}

\item{fontsize}{an integer for font size of plot}

\item{angle.names}{an integer to turn the names of plot}

\item{big}{use it for a better visualization for high gen parameter (TRUE or FALSE)}

\item{Amatrix}{relationship matrix in same order of the pedigree (matrix)}
}
\value{
Graphic with ind genealogy until gen generations ago
}
\description{
Builds a genealogy for a given individual based on his pedigree
}
\examples{
data(ped.mrode)
#Build Amatrix diploid (no double reduction proportion)
Amat <- Amatrix(data=ped.mrode,ploidy=2,unk=0)
#Build Amatrix autotetraploidy (double reduction proportion=0.1)
anc.plot(data=ped.mrode,ind="Var4",Amatrix=Amat)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}

