% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.potato-data.R
\docType{data}
\name{snp.potato}
\alias{snp.potato}
\title{Autotetraploid Molecular Data}
\format{matrix}
\usage{
data(snp.potato)
}
\description{
SNP marker matrix from Autotetraploid potato with dosage information coded as 0, 1, 2, 3, 4. Derived from the Supplementary data (FileS2.csv) from Endelman (2018).
}
\examples{
data(snp.potato)

}
\references{
Endelman, J. B., et al., 2018. Genetic variance partitioning and genome-wide prediction with allele dosage information in autotetraploid potato. Genetics, 209(1) pp. 77-87.
}
\keyword{datasets}
