% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.sol-data.R
\docType{data}
\name{snp.sol}
\alias{snp.sol}
\title{Molecular data for autopolyploid examples}
\format{data.frame}
\usage{
data(snp.sol)
}
\description{
Dataset extract from  supplementary material from Endelman et al. (2018). 
SNP marker matrix from Pine tree coded as 0,1,2,3,4 and missing value as -9.
}
\examples{
data(snp.sol)

}
\references{
Endelman, JB, et al., 2018 Genetic variance partitioning and genome-wide prediction with allele dosage information in autotetraploid potato. Genetics, 209(1) pp. 77-87.
}
\keyword{datasets}
