% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.table-data.R
\docType{data}
\name{snp.table}
\alias{snp.table}
\title{Molecular Data}
\format{table}
\usage{
data(snp.table)
}
\description{
Dummy data for construction of relationship matrix based on SNPs. Matrix (markers x genotype). Dosage as 0,1, and 2, and missing value as -9.
}
\examples{
data(snp.table)

}
\keyword{datasets}
