% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortped.R
\name{sortped}
\alias{sortped}
\title{Organizes pedigree data in a chronological way}
\usage{
sortped(data = NULL, loop.in = 1000, loop.between = 100, print = FALSE)
}
\arguments{
\item{data}{name of the pedigree data. Default=NULL.}

\item{loop.in}{max number of iteractions to get the chronological order inside a column (e.g. parental 1 or parental 2). Default = 1000}

\item{loop.between}{max number of iteractions to get the chronological order between columns. Default=100.}

\item{print}{if TRUE print the step-by-step of the loops. Default=FALSE.}
}
\value{
list with parental 1, parental 2, and real names of the individuals (key) also saves a txt file with the new chronological pedigree.
}
\description{
This function organizes pedigree data in a chronological way and return 3 lists: i) parental 1 values (numeric); ii) parental 2 values (numeric); iii) real names of the individuals. Also save a .txt file with new pedigree file.
}
\examples{
data.treat()

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}

