% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyped.R
\name{verifyped}
\alias{verifyped}
\title{Survying on verify pedigree}
\usage{
verifyped(pedigree, unk = 0)
}
\arguments{
\item{pedigree}{data name from a pedigree list}

\item{unk}{unknown value of your data}
}
\value{
i) warnings with the conflicting data regarding number of columns in the data, repeated entries, missing parental names in the entries. ii) TRUE if there was any warning.
}
\description{
This function verify which rows in a pedigree data has missing parental or conflictuos data
}
\examples{
verifyped()

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}

