% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amatrix.R
\name{Amatrix}
\alias{Amatrix}
\title{Construction of Relationship Matrix A}
\usage{
Amatrix(data = NULL, ploidy = 2, w = 0, unk = 0, verify = TRUE)
}
\arguments{
\item{data}{pedigree data name (3-column way format).}

\item{ploidy}{2 or 4 (default=2).}

\item{w}{proportion of parental gametas IBD due to double reduction (default=0).}

\item{unk}{string or number indicating the unknown value related in the pedigree file (default=0).}

\item{verify}{verifies pedigree file for conflictuos entries (default=TRUE).}
}
\value{
Matrix with the Relationship between the individuals.
}
\description{
Creates a Relationship Matrix A from a pedigree data in a 3-column way format based on ploidy level (2 or 4) and, if ploidy equals 4, based on proportion of parental gametes that are IBD (Identical by Descent) due to double reduction.
}
\examples{
data(ped.mrode)
#Build Amatrix diploid (no double reduction proportion)
Amatrix(data=ped.mrode,ploidy=2,unk=0)
#Build Amatrix autotetraploidy (double reduction proportion=0.1)
Amatrix(data=ped.mrode,ploidy=4,w=0.1,unk=0)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
\references{
\emph{Chapter 2: Genetic Covariance Between Relatives in Mrode, R. A., and Thompson, R. Linear models for the prediction of animal breeding values. Cabi, 2005.}

\emph{Slater, A. T., Wilson, G. M., Cogan, N. O., Forster, J. W., & Hayes, B. J. (2013). Improving the analysis of low heritability complex traits for enhanced genetic gain in potato. Theoretical and Applied Genetics, 1-12.}
}

