% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciitonumber.R
\name{asciitonumber}
\alias{asciitonumber}
\title{Given a pedigree ascii format returns a numeric list}
\usage{
asciitonumber(pedigree.data, unk = 0)
}
\arguments{
\item{pedigree.data}{matrix with 3 columns (without row names/column names), the first for the individual name, second for parent 1, third for parent 2.}

\item{unk}{code of the data missing, default = 0}

\item{force}{replace all the missing parents with the unk value}
}
\value{
List with $sire: vector with the numbers respectives to the first column parent in the data; $dire: as sire but for second column; $ind.data: the matrix of original data.
}
\description{
This function creates a list with numeric indices given a pedigree data. Also it checks if all the listed parent name are listed before in the individual name column and if the parent exist in the matrix .
}
\examples{
asciitonumber()

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}

