% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gmatrix.R
\name{Gmatrix}
\alias{Gmatrix}
\title{Construction of Relationship Matrix G}
\usage{
Gmatrix(SNPmatrix = NULL, method = 2, missingValue = -9, maf = 0,
  verify.posdef = TRUE)
}
\arguments{
\item{SNPmatrix}{matrix (n x m), where n is is individual names and m is marker names (coded inside the matrix as 0,1,2,missingValue).}

\item{method}{"Powell" (1), "VanRaden" (2), shared markers between individual (3). Default=2.}

\item{missingValue}{missing value in data. Default=-9.}

\item{maf}{max of missing data accepted to each marker. Default=0.}

\item{verify.posdef}{verify if the resulting matrix is positive-definite. Default=TRUE.}
}
\value{
Matrix with the Relationship between the individuals
}
\description{
Given a matrix (individual x markers), a method ((1) Powell or (2) VanRaden), a missing value, and a maf threshold, return a G matrix. Also the function can calculate the number of markers shared two-by-two individuals (method 3).
}
\examples{
data(snp.pine)
#Verifying if data is coded as 0,1,2 and missing value.
str(snp.pine)
#Build Gmatrix
PineGmatrix <- Gmatrix(snp.pine,missingValue=-9,maf=0.05)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com} and Marcio Resende Jr.
}
\references{
\emph{Van Raden, P. M. (2008). Efficient methods to compute genomic predictions. Journal of Dairy Science, 91(11), 4414-4423.}

\emph{Powell, J. E. (2010). Reconciling the analysis of IBD and IBS in complex trait studies. Nature Reviews Genetics, 11(11), 800-805.}
}

