% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.pine-data.R
\docType{data}
\name{snp.pine}
\alias{snp.pine}
\title{Molecular data for diploid examples}
\format{matrix}
\usage{
data(snp.pine)
}
\description{
Dataset extract from  supplementary material from Resende et al. (2012). 
SNP marker matrix from Pine tree coded as 0,1, and 2, and missing value as -9.
}
\examples{
data(snp.pine)

}
\references{
Resende, MF, et al., 2012 Accuracy of genomic selection methods in a standard data set of loblolly pine (Pinus taeda l.). Genetics 190: 1503–1510.
}
\keyword{datasets}
